2 hours on error for declaring outside a callback
Because I was trying to declare a variable by getting element by ID OUTSIDE the callback function,
it returned "None" on the unwrap. This is not even a crazy problem, but the framework and the fact
this was written in Rust made it unbelievably difficult to track down. I need more fundamentals of
component frameworks, first of all. The worry about the ways that Rust is weird also pushed into wrong
corners. Not knowing how to deal with moves/clones led to gravitating into the wrong direction,
into ways of breaking a component-framework render system... Difficult stuff, mixing Rust and Yew, without
having something like React 'fully' mastered.